refactor(routes/controller): separate routes
Separate the controller and routes in their own
    file
